Separate execution context for ColumnStore read paths
This is to prevent a deadlock when appending segments, where getSegFromCache issues a read future;
if it uses the same write context which is blocking for backpressure then the write will deadlock
waiting for the read future which is waiting for other futures.
