+++ core/src/test/scala/filodb.core/store/InMemoryColumnStoreSpec.scala
@@ -2,5 +2,5 @@ package filodb.core.store
 
 class InMemoryColumnStoreSpec extends ColumnStoreSpec {
   import scala.concurrent.ExecutionContext.Implicits.global
-  val colStore = new InMemoryColumnStore
+  val colStore = new InMemoryColumnStore(global)
 }
